app/requirements.txt
pypdf
langchain-core
langchain-openai
OpenAI 
langchain-community
python-docx
fastapi[standard]